rename tests folder functional->acceptance
