+++ java-specific/src/test/java/com/pryv/acceptance/ConnectionTest.java
@@ -1,4 +1,4 @@
-package com.pryv.functional;
+package com.pryv.acceptance;
 
 
 import com.jayway.awaitility.Awaitility;
